Module aws_sdk_resourcegroups::error

Modules

bad_request_exception

See BadRequestException

forbidden_exception

See ForbiddenException

internal_server_error_exception

See InternalServerErrorException

method_not_allowed_exception

See MethodNotAllowedException

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

unauthorized_exception

See UnauthorizedException

Structs

BadRequestException

The request includes one or more parameters that violate validation rules.

CreateGroupError

DeleteGroupError

ForbiddenException

The caller isn't authorized to make the request. Check permissions.

GetGroupConfigurationError

GetGroupError

GetGroupQueryError

GetTagsError

GroupResourcesError

InternalServerErrorException

An internal error occurred while processing the request. Try again later.

ListGroupResourcesError

ListGroupsError

MethodNotAllowedException

The request uses an HTTP method that isn't allowed for the specified resource.

NotFoundException

One or more of the specified resources don't exist.

PutGroupConfigurationError

SearchResourcesError

TagError

TooManyRequestsException

You've exceeded throttling limits by making too many requests in a period of

UnauthorizedException

The request was rejected because it doesn't have valid credentials for the target

UngroupResourcesError

UntagError

UpdateGroupError

UpdateGroupQueryError

Enums

CreateGroupErrorKind

DeleteGroupErrorKind

GetGroupConfigurationErrorKind

GetGroupErrorKind

GetGroupQueryErrorKind

GetTagsErrorKind

GroupResourcesErrorKind

ListGroupResourcesErrorKind

ListGroupsErrorKind

PutGroupConfigurationErrorKind

SearchResourcesErrorKind

TagErrorKind

UngroupResourcesErrorKind

UntagErrorKind

UpdateGroupErrorKind

UpdateGroupQueryErrorKind