Module aws_sdk_resourcegroups::model

Modules

failed_resource

See FailedResource

group

See Group

group_configuration

See GroupConfiguration

group_configuration_item

See GroupConfigurationItem

group_configuration_parameter

See GroupConfigurationParameter

group_filter

See GroupFilter

group_identifier

See GroupIdentifier

group_query

See GroupQuery

list_group_resources_item

See ListGroupResourcesItem

pending_resource

See PendingResource

query_error

See QueryError

resource_filter

See ResourceFilter

resource_identifier

See ResourceIdentifier

resource_query

See ResourceQuery

resource_status

See ResourceStatus

Structs

FailedResource

A resource that failed to be added to or removed from a group.

Group

A resource group that contains AWS resources. You can assign resources to the group

GroupConfiguration

A service configuration associated with a resource group. The configuration options

GroupConfigurationItem

An item in a group configuration. A group service configuration can have one or more

GroupConfigurationParameter

A parameter for a group configuration item. For details about group service

GroupFilter

A filter collection that you can use to restrict the results from a List

GroupIdentifier

The unique identifiers for a resource group.

GroupQuery

A mapping of a query attached to a resource group that determines the AWS resources

ListGroupResourcesItem

A structure returned by the ListGroupResources operation that

PendingResource

A structure that identifies a resource that is currently pending addition to the group

QueryError

A two-part error structure that can occur in ListGroupResources or

ResourceFilter

A filter name and value pair that is used to obtain more specific results from a list

ResourceIdentifier

A structure that contains the ARN of a resource and its resource type.

ResourceQuery

The query that is used to define a resource group or a search for resources. A query

ResourceStatus

A structure that identifies the current group membership status for a resource. Adding

Enums

GroupConfigurationStatus

GroupFilterName

QueryErrorCode

QueryType

ResourceFilterName

ResourceStatusValue