Module aws_sdk_resourcegroups::operation

Structs

CreateGroup

Creates a resource group with the specified name and description. You can optionally

DeleteGroup

Deletes the specified resource group. Deleting a resource group does not delete any

GetGroup

Returns information about a specified resource group.

GetGroupConfiguration

Returns the service configuration associated with the specified resource group. For

GetGroupQuery

Retrieves the resource query associated with the specified resource group. For more

GetTags

Returns a list of tags that are associated with a resource group, specified by an

GroupResources

Adds the specified resources to the specified group.

ListGroupResources

Returns a list of ARNs of the resources that are members of a specified resource

ListGroups

Returns a list of existing resource groups in your account.

PutGroupConfiguration

Attaches a service configuration to the specified group. This occurs asynchronously,

SearchResources

Returns a list of AWS resource identifiers that matches the specified query. The

Tag

Adds tags to a resource group with the specified ARN. Existing tags on a resource

UngroupResources

Removes the specified resources from the specified group.

Untag

Deletes tags from a specified resource group.

UpdateGroup

Updates the description for an existing group. You cannot update the name of a

UpdateGroupQuery

Updates the resource query of a group. For more information about resource queries,